# Faster time parsing (~93% faster)