# Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and enhancements and this helps us generate change logs for our releases. You can link an issue to this PR using the GitHub syntax. For example `Closes #123` indicates that this PR will close issue #123.
-->

Closes #.

# Rationale for this change
 
<!--
Why are you proposing this change? If this is already explained clearly in the issue then this section is not needed.
Explaining clearly why changes are proposed helps reviewers understand your changes and offer better suggestions for fixes.
-->

Faster time parsing

```
9:50                    time:   [5.1079 ns 5.1315 ns 5.1503 ns]
                        change: [-92.180% -92.146% -92.111%] (p = 0.00 < 0.05)
                        Performance has improved.

09:50                   time:   [5.1957 ns 5.2014 ns 5.2086 ns]
                        change: [-92.186% -92.139% -92.077%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 19 outliers among 100 measurements (19.00%)
  1 (1.00%) low mild
  2 (2.00%) high mild
  16 (16.00%) high severe

09:50 PM                time:   [6.0355 ns 6.0371 ns 6.0387 ns]
                        change: [-93.754% -93.746% -93.739%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  5 (5.00%) high mild
  3 (3.00%) high severe

9:50:12 AM              time:   [6.0481 ns 6.0591 ns 6.0733 ns]
                        change: [-95.110% -95.105% -95.098%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 11 outliers among 100 measurements (11.00%)
  3 (3.00%) high mild
  8 (8.00%) high severe

09:50:12 PM             time:   [6.4839 ns 6.4934 ns 6.5036 ns]
                        change: [-94.865% -94.853% -94.838%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 14 outliers among 100 measurements (14.00%)
  6 (6.00%) high mild
  8 (8.00%) high severe

09:50:12.123456789      time:   [10.475 ns 10.503 ns 10.526 ns]
                        change: [-92.510% -92.485% -92.459%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 27 outliers among 100 measurements (27.00%)
  23 (23.00%) low severe
  1 (1.00%) low mild
  1 (1.00%) high mild
  2 (2.00%) high severe

9:50:12.123456789       time:   [9.9941 ns 10.036 ns 10.082 ns]
                        change: [-92.728% -92.701% -92.673%] (p = 0.00 < 0.05)
                        Performance has improved.

09:50:12.123456789 PM   time:   [11.016 ns 11.037 ns 11.060 ns]
                        change: [-93.423% -93.410% -93.396%] (p = 0.00 < 0.05)
                        Performance has improved.
```

# What changes are included in this PR?

<!--
There is no need to duplicate the description in the issue here but it is sometimes worth providing a summary of the individual changes in this PR.
-->

# Are there any user-facing changes?


<!--
If there are user-facing changes then we may require documentation to be updated before approving the PR.
-->

<!---
If there are any breaking changes to public APIs, please add the `breaking change` label.
-->